♻️ Move connector.go to internal/wrap
